requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
mistune==3.0.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
smmap==5.0.1
soupsieve==2.6
urllib3==2.2.2